requirements.txt
FastAPI==0.71.0
Uvicorn==0.17.0
Alembic==1.7.5
psycopg2==2.9.3
python-jose==3.3.0
passlib==1.7.4
pydantic==1.9.0
numpy==1.22.1
python-multipart==0.0.5
PyYAML==6.0
pytest==7.0.1
requests==2.27.1
